tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-delay_one_value.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-delay value with one list entry</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-delay-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="1s is a valid animation-delay value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow {
         from {
            width: 100px;
         }
         to {
            width: 200px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         animation-duration: 1s;
         animation-delay: 1s;
         animation-fill-mode: forwards;
      }
      .test {
         animation-name: grow;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxDeviation = 100;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function() {
            clearTimeout(timeout);
            var delaySeen = Date.now() - delayStart;
            var passed = aboutEqual(delaySeen, 1000);
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.'
                               : 'FAIL (got ' + delaySeen + ')');
         });
         var delayStart = Date.now();
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should start increasing in width after one
   second.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
